+++ src/generate/index.ts
@@ -2,10 +2,10 @@ import { CommandModule } from "yargs";
 import * as chalk from "chalk";
 import * as path from "path";
 import * as fs from "fs";
-import { logger } from "../log";
-import { directoryExists, findNodeCGDirectory, getNodeCGIODirectory } from "../fsUtils";
-import { readInstallInfo } from "../installation";
-import { corePackages } from "../install/nodecgIOVersions";
+import { logger } from "../utils/log";
+import { directoryExists, findNodeCGDirectory, getNodeCGIODirectory } from "../utils/fs";
+import { readInstallInfo } from "../utils/installation";
+import { corePackages } from "../nodecgIOVersions";
 import { GenerationOptions, promptGenerationOpts } from "./prompt";
 import * as defaultTsConfigJson from "./tsconfig.json";
 import CodeBlockWriter from "code-block-writer";
