+++ src/index.ts
@@ -2,7 +2,7 @@ import * as yargs from "yargs";
 import { installModule } from "./install";
 import { uninstallModule } from "./uninstall";
 import { version } from "../package.json";
-import { checkForCliUpdate, ensureNode12 } from "./cli";
+import { checkForCliUpdate, ensureNode12 } from "./utils/cli";
 import { generateModule } from "./generate";
 
 // This file gets imported by the index.js file of the repository root.
