+++ src/install/index.ts
@@ -1,14 +1,14 @@
 import { CommandModule } from "yargs";
 import * as path from "path";
-import { directoryExists, findNodeCGDirectory, getNodeCGIODirectory, removeDirectory } from "../fsUtils";
+import { directoryExists, findNodeCGDirectory, getNodeCGIODirectory, removeDirectory } from "../utils/fs";
 import { createDevInstall } from "./development";
-import { manageBundleDir } from "../nodecgConfig";
+import { manageBundleDir } from "../utils/nodecgConfig";
 import { promptForInstallInfo } from "./prompt";
-import { readInstallInfo } from "../installation";
+import { readInstallInfo } from "../utils/installation";
 import { createProductionInstall } from "./production";
 import * as os from "os";
-import { logger } from "../log";
-import { requireNpmV7 } from "../npm";
+import { logger } from "../utils/log";
+import { requireNpmV7 } from "../utils/npm";
 
 export const installModule: CommandModule<unknown, { concurrency: number }> = {
     command: "install",
