+++ src/install/prompt.ts
@@ -1,8 +1,8 @@
-import { Installation, ProductionInstallation } from "../installation";
+import { Installation, ProductionInstallation } from "../utils/installation";
 import * as inquirer from "inquirer";
-import { getHighestPatchVersion, getMinorVersions, NpmPackage } from "../npm";
+import { getHighestPatchVersion, getMinorVersions, NpmPackage } from "../utils/npm";
 import * as semver from "semver";
-import { logger } from "../log";
+import { logger } from "../utils/log";
 import {
     corePackage,
     corePackages,
@@ -11,7 +11,7 @@ import {
     developmentVersion,
     getServicesForVersion,
     supportedNodeCGIORange,
-} from "./nodecgIOVersions";
+} from "../nodecgIOVersions";
 
 interface PromptInput {
     version: string;
