+++ test/utils/fs.ts
@@ -1,9 +1,15 @@
 import { createFsFromVolume, vol } from "memfs";
-import { directoryExists, ensureDirectory, executeCommand, findNodeCGDirectory, removeDirectory } from "../src/fsUtils";
+import {
+    directoryExists,
+    ensureDirectory,
+    executeCommand,
+    findNodeCGDirectory,
+    removeDirectory,
+} from "../../src/utils/fs";
 import * as path from "path";
 import * as child_process from "child_process";
-import { testDir } from "./testUtils";
-import { logger } from "../src/log";
+import { testDir } from "../testUtils";
+import { logger } from "../../src/utils/log";
 
 jest.mock("fs", () => createFsFromVolume(vol));
 afterEach(() => vol.reset());
